Move owned-dialog-specific tests from the general suite to its own
