+++ composite/timeline/src/test/java/br/com/orcinus/orca/composite/timeline/search/field/OwnedResultSearchTextFieldDialogTests.kt
@@ -19,30 +19,95 @@ import android.view.View
 import android.view.Window
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.compose.ui.test.junit4.createAndroidComposeRule
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.ViewModelStoreOwner
 import androidx.lifecycle.findViewTreeLifecycleOwner
 import androidx.lifecycle.findViewTreeViewModelStoreOwner
+import androidx.lifecycle.setViewTreeLifecycleOwner
+import androidx.lifecycle.setViewTreeViewModelStoreOwner
+import androidx.savedstate.SavedStateRegistryOwner
 import androidx.savedstate.findViewTreeSavedStateRegistryOwner
-import androidx.test.core.app.launchActivity
+import androidx.savedstate.setViewTreeSavedStateRegistryOwner
 import assertk.all
 import assertk.assertThat
 import assertk.assertions.isNotNull
+import assertk.assertions.isSameInstanceAs
 import assertk.assertions.prop
+import io.mockk.mockk
 import kotlin.test.Test
+import org.junit.Rule
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 
 @RunWith(RobolectricTestRunner::class)
 internal class OwnedResultSearchTextFieldDialogTests {
+  private inline val activity
+    get() = composeRule.activity
+
+  @get:Rule val composeRule = createAndroidComposeRule<ComponentActivity>()
+
+  @Test
+  fun doesNotDeconfigureActivityViewTreeOwnershipWhenDismissedWhileNotShown() {
+    val originalViewTreeLifecycleOwner = mockk<LifecycleOwner>()
+    val originalViewTreeViewModelStoreOwner = mockk<ViewModelStoreOwner>()
+    val originalViewTreeSavedStateRegistryOwner = mockk<SavedStateRegistryOwner>()
+    activity.window?.decorView?.apply {
+      setViewTreeLifecycleOwner(originalViewTreeLifecycleOwner)
+      setViewTreeViewModelStoreOwner(originalViewTreeViewModelStoreOwner)
+      setViewTreeSavedStateRegistryOwner(originalViewTreeSavedStateRegistryOwner)
+    }
+    OwnedResultSearchTextFieldDialog(activity).dismiss()
+    assertThat(activity).prop(ComponentActivity::getWindow).prop(Window::getDecorView).all {
+      prop(View::findViewTreeLifecycleOwner).isSameInstanceAs(originalViewTreeLifecycleOwner)
+      prop(View::findViewTreeViewModelStoreOwner)
+        .isSameInstanceAs(originalViewTreeViewModelStoreOwner)
+      prop(View::findViewTreeSavedStateRegistryOwner)
+        .isSameInstanceAs(originalViewTreeSavedStateRegistryOwner)
+    }
+  }
+
+  @Test
+  fun deconfiguresActivityViewTreeOwnershipWhenDismissed() {
+    val originalViewTreeLifecycleOwner = mockk<LifecycleOwner>()
+    val originalViewTreeViewModelStoreOwner = mockk<ViewModelStoreOwner>()
+    val originalViewTreeSavedStateRegistryOwner = mockk<SavedStateRegistryOwner>()
+    activity.window?.decorView?.apply {
+      setViewTreeLifecycleOwner(originalViewTreeLifecycleOwner)
+      setViewTreeViewModelStoreOwner(originalViewTreeViewModelStoreOwner)
+      setViewTreeSavedStateRegistryOwner(originalViewTreeSavedStateRegistryOwner)
+    }
+    OwnedResultSearchTextFieldDialog(activity)
+      .apply(OwnedResultSearchTextFieldDialog::show)
+      .dismiss()
+    assertThat(activity).prop(ComponentActivity::getWindow).prop(Window::getDecorView).all {
+      prop(View::findViewTreeLifecycleOwner).isSameInstanceAs(originalViewTreeLifecycleOwner)
+      prop(View::findViewTreeViewModelStoreOwner)
+        .isSameInstanceAs(originalViewTreeViewModelStoreOwner)
+      prop(View::findViewTreeSavedStateRegistryOwner)
+        .isSameInstanceAs(originalViewTreeSavedStateRegistryOwner)
+    }
+  }
+
+  @Test
+  fun initializesActivityViewTreeOwnersOnlyOnceWhenShown() {
+    OwnedResultSearchTextFieldDialog(activity).apply(OwnedResultSearchTextFieldDialog::show).also {
+      assertThat(activity).prop(ComponentActivity::getWindow).prop(Window::getDecorView).all {
+        prop(View::findViewTreeLifecycleOwner).isSameInstanceAs(composeRule.activity)
+        prop(View::findViewTreeViewModelStoreOwner).isSameInstanceAs(composeRule.activity)
+        prop(View::findViewTreeSavedStateRegistryOwner).isSameInstanceAs(composeRule.activity)
+      }
+    }
+  }
+
   @Test
   fun configuresActivityViewTreeOwnershipWhenShown() {
-    launchActivity<ComponentActivity>().onActivity {
-      val dialog = OwnedResultSearchTextFieldDialog(it)
-      it.setContent { dialog.Content() }
-      assertThat(it).prop(ComponentActivity::getWindow).prop(Window::getDecorView).all {
-        prop(View::findViewTreeLifecycleOwner).isNotNull()
-        prop(View::findViewTreeViewModelStoreOwner).isNotNull()
-        prop(View::findViewTreeSavedStateRegistryOwner).isNotNull()
-      }
+    val dialog = OwnedResultSearchTextFieldDialog(activity)
+    activity.setContent { dialog.Content() }
+    assertThat(activity).prop(ComponentActivity::getWindow).prop(Window::getDecorView).all {
+      prop(View::findViewTreeLifecycleOwner).isNotNull()
+      prop(View::findViewTreeViewModelStoreOwner).isNotNull()
+      prop(View::findViewTreeSavedStateRegistryOwner).isNotNull()
     }
   }
 }
